requirements.txt
Django==2.0.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
PyJWT==1.5.3
pytz==2017.3
